6th task

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,11 @@ int main(void)
 	long int l;
 	long long int ll;
 	float f;
-	printf("size of a char: %lu byte(s)\n", (unsigned long) sizeof(char));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long) sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long) sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", (signed long) sizeof(long long int));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long) sizeof(float));
+	
+	printf("size of a char: %lu byte(s)\n", sizeof(char));
+	printf("Size of an int: %lu byte(s)\n", sizeof(int));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
+	printf("Size of a float: %lu byte(s)\n", sizeof(float));
 	return (0);
 }
